+++ mmseg/structures/__init__.py
@@ -1,18 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .amg import (MaskData, area_from_rle, batch_iterator, batched_mask_to_box,
-                  box_xyxy_to_xywh, build_all_layer_point_grids,
-                  calculate_stability_score, coco_encode_rle,
-                  generate_crop_boxes, is_box_near_crop_edge,
-                  mask_to_rle_pytorch, remove_small_regions, rle_to_mask,
-                  uncrop_boxes_xyxy, uncrop_masks, uncrop_points)
 from .sampler import BasePixelSampler, OHEMPixelSampler, build_pixel_sampler
 from .seg_data_sample import SegDataSample
 
 __all__ = [
     'SegDataSample', 'BasePixelSampler', 'OHEMPixelSampler',
-    'build_pixel_sampler', 'MaskData', 'area_from_rle', 'batch_iterator',
-    'batched_mask_to_box', 'box_xyxy_to_xywh', 'build_all_layer_point_grids',
-    'calculate_stability_score', 'coco_encode_rle', 'generate_crop_boxes',
-    'is_box_near_crop_edge', 'mask_to_rle_pytorch', 'remove_small_regions',
-    'rle_to_mask', 'uncrop_boxes_xyxy', 'uncrop_masks', 'uncrop_points'
+    'build_pixel_sampler'
 ]
