+++ projects/sam_inference_demo/sam/sam_inferencer.py
@@ -4,18 +4,17 @@ from typing import Any, Dict, List, Optional, Tuple
 import numpy as np
 import torch
 from mmengine.runner.checkpoint import load_checkpoint
+# yapf: disable
+from sam.utils import (MaskData, area_from_rle, batch_iterator,
+                       batched_mask_to_box, box_xyxy_to_xywh,
+                       build_all_layer_point_grids, calculate_stability_score,
+                       coco_encode_rle, generate_crop_boxes,
+                       is_box_near_crop_edge, mask_to_rle_pytorch,
+                       remove_small_regions, rle_to_mask, uncrop_boxes_xyxy,
+                       uncrop_masks, uncrop_points)
 from torchvision.ops.boxes import batched_nms, box_area
 
 from mmseg.registry import MODELS, TRANSFORMS
-# yapf: disable
-from mmseg.structures.amg import (MaskData, area_from_rle, batch_iterator,
-                                  batched_mask_to_box, box_xyxy_to_xywh,
-                                  build_all_layer_point_grids,
-                                  calculate_stability_score, coco_encode_rle,
-                                  generate_crop_boxes, is_box_near_crop_edge,
-                                  mask_to_rle_pytorch, remove_small_regions,
-                                  rle_to_mask, uncrop_boxes_xyxy, uncrop_masks,
-                                  uncrop_points)
 
 # yapf: enable
